Add 'toggleSupplier' function to capture logic
